used SonarQube to fix code issues in card & charms

+++ core/src/main/java/io/wasabi/urg/elements/card/ExtraChange.java
@@ -15,8 +15,8 @@ public class ExtraChange extends Card {
 
     @Override
     public void roundStartEffect() {
-        //System.out.println("ExtraChange card effect triggered");
+
         triggerDisplay();
-        runState.addChips((int) Roulette.getInstance().getRoundManager().getCurrentConfig().getQuota()/10);
+        runState.addChips(Roulette.getInstance().getRoundManager().getCurrentConfig().getQuota()/10);
     }
 }
